refactor(#141): split terrain.py and aerodynamics.py into focused modules
Extract cohesive submodules from two of the >1000 LOC monoliths flagged by
issue #141, keeping the public API intact via re-exports in the original
module (so `from src.shared.python.physics.terrain import ...` and
`from src.shared.python.physics.aerodynamics import ...` keep working).

terrain.py (1199 -> 468 LOC) now re-exports from three new submodules:
  - terrain_materials.py (266): TerrainType, SurfaceMaterial, MATERIALS,
    TERRAIN_MATERIAL_MAP
  - terrain_elevation.py (339): ElevationMap (height grid, bilinear
    interpolation, gradient / normal / slope-angle queries)
  - terrain_regions.py   (225): TerrainPatch, TerrainRegion (rect + complex
    shapes with containment tests and dict serialization)

aerodynamics.py (1091 -> 260 LOC) now re-exports from four new submodules:
  - aerodynamics_config.py      (167): AerodynamicsConfig, WindConfig,
    RandomizationConfig
  - aerodynamics_forces.py      (311): DragModel, LiftModel, MagnusModel
  - aerodynamics_wind.py        (288): WindGust, TurbulenceModel, WindModel
  - aerodynamics_environment.py (174): EnvironmentSnapshot,
    EnvironmentRandomizer

All resulting files <=500 LOC. Verified with:
  - `ruff check` clean on every new/modified file
  - `ruff format --check` clean on every new/modified file
  - `pytest tests/unit/test_terrain.py tests/unit/test_aerodynamics.py`
    -> 127 passed
  - `pytest tests/unit/test_terrain_engine.py
    tests/unit/test_enhanced_ball_flight.py
    tests/integration/test_terrain_cross_engine.py` -> 47 passed,
    24 skipped (unrelated, Rust kernel missing)

No test logic touched; no import sites needed to change because every
existing caller uses `from src.shared.python.physics.<module> import ...`
and the public names are preserved.

Partially closes #141 (2 of 15 monoliths; mesh_generator.py tracked in #136).
